Devons branch

+++ src/main/java/spring2024/cs472/hotelwebsite/controllers/AdminAccountCrudController.java
@@ -1,5 +1,6 @@
 package spring2024.cs472.hotelwebsite.controllers;
 
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -20,12 +21,18 @@ public class AdminAccountCrudController {
     public AdminAccountCrudController(AccountService accountService) {this.accountService = accountService;}
 
     @GetMapping("/AccountIndex")
-    public String showUserList(Model model) {
+    public String showUserList(Model model, HttpSession session) {
+        if(session.getAttribute("admin") == null){
+            return "redirect:/login";
+        }
         model.addAttribute("Admins", accountService.getAllAdmins());
         return "accountIndex";
     }
     @GetMapping("/admin/edit/{id}")
-    public String showUpdateForm(@PathVariable("id") long id, Model model) {
+    public String showUpdateForm(@PathVariable("id") long id, Model model, HttpSession session) {
+        if(session.getAttribute("admin") == null){
+            return "redirect:/login";
+        }
         Admin admin = (Admin) accountRepository.findById(id);
 //                .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
 
@@ -34,7 +41,10 @@ public class AdminAccountCrudController {
     }
     @PostMapping("/admin/edit/{id}")
     public String updateAdmin(@PathVariable("id") long id, Admin admin,
-                              BindingResult result) {
+                              BindingResult result, HttpSession session) {
+        if(session.getAttribute("admin") == null){
+            return "redirect:/login";
+        }
         if (result.hasErrors()) {
             return "updateAdmin";
         }
@@ -43,7 +53,10 @@ public class AdminAccountCrudController {
         return "redirect:/AccountIndex";
     }
     @GetMapping("/admin/delete/{id}")
-    public String deleteAdmin(@PathVariable("id") long id, Model model) {
+    public String deleteAdmin(@PathVariable("id") long id, Model model, HttpSession session) {
+        if(session.getAttribute("admin") == null){
+            return "redirect:/login";
+        }
         Admin admin = (Admin) accountRepository.findById(id);
 //                .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
         accountRepository.delete(admin);
